[#115] Move safecall to lib. Update copyright date
