+++ app/services/server_access/heroku.rb
@@ -91,7 +91,7 @@ module ServerAccess
     end
 
     def safe_call
-      @_safe_call ||= ::ServerAccess::HerokuHelpers::SafeCall.new(
+      @_safe_call ||= Utils::SafeCall.new(
         exceptions: [Excon::Error::UnprocessableEntity, Excon::Error::NotFound],
         errors_processor: ->(error) { error.response.data[:body] }
       )
