+++ app/views/landings/index.html.erb
@@ -294,7 +294,7 @@
       <div class="footer_item-copy">
         <img src="static/images/logo.svg" alt>
         <small class="footer_copyright">
-          ©2019 DeployQA. All rights reserved.
+          ©2020 DeployQA. All rights reserved.
         </small>
       </div>
